requirements.txt
gradio==4.36.1
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.1
langchain-huggingface==0.1.0
sentence-transformers==3.1.1
faiss-cpu==1.12.0
python-dotenv==1.0.1
openai==1.54.0
httpx==0.27.2

